feat: move logic to new route
