+++ src/routes/+page.svelte
@@ -1,114 +1,21 @@
-<script lang="ts">
-  import IssueCard from '$lib/components/issue-card.svelte';
-  import Search from '$lib/components/search.svelte';
-  import Filter from '$lib/components/filter.svelte';
-  import LoadMore from '$lib/components/load-more.svelte';
-  import Toggle from '$lib/components/toggle.svelte';
-  import Seo from '$lib/components/seo.svelte';
-
-  import { selectedLabels } from '$lib/stores/selected-labels.store';
-  import type { SearchResponse } from '../global';
-  import { goto } from '$app/navigation';
-  import type { PageData } from './$types';
-  import { query } from '$lib/constants.json';
-  import { navigating } from '$app/stores';
-  import Loader from '$lib/components/loader.svelte';
-  export let data: PageData;
-
-  let { checked } = data;
-  let loadDisabled = false;
-  $: githubData = data.data;
-
-  $: searchString = '';
-  $: filteredLabels = githubData.edges;
-  $: searchItems = githubData.edges;
-
-  $: $selectedLabels, filterLabels();
-  const filterLabels = () => {
-    filteredLabels = githubData.edges.filter((githubDataset) =>
-      $selectedLabels.every((label) =>
-        githubDataset.node.labels.edges.some((edge) => edge.node.name === label),
-      ),
-    );
-  };
-
-  const fetchMore = async () => {
-    loadDisabled = true;
-    const res = await fetch('/api/get-issues', {
-      method: 'POST',
-      body: JSON.stringify({
-        after: githubData.pageInfo.endCursor,
-        query: checked ? query.global : query.org,
-      }),
-    });
-    if (res.ok) {
-      const respData = (await res.json()) as SearchResponse;
-      githubData.edges = [...githubData.edges, ...respData.edges];
-      githubData.pageInfo = respData.pageInfo;
-      const uniqueLabels = [...githubData.labels, ...respData.labels];
-      githubData.labels = [...new Set(uniqueLabels)];
-    }
-    loadDisabled = false;
-  };
-
-  const onChangeHandler = async () => {
-    if (checked) {
-      await goto('?global=true');
-      performSearch();
-      return;
-    }
-    await goto('?global=false');
-    performSearch();
-  };
-
-  const performSearch = () => {
-    searchItems = githubData.edges.filter((el) =>
-      el.node.title.toLowerCase().includes(searchString.toLowerCase()),
-    );
-  };
-
-  $: intersectedArray = filteredLabels.filter((item) => searchItems.includes(item));
-
-  // SEO Parameters
-  const title = 'Good First Issue Finder by EddieHub';
-  const metadescription =
-    'Good First Issue Finder helps new open source contributors pave their path into the world of open source through good first issues.';
+<script>
+  import Link from '$lib/components/index/link.svelte';
 </script>
 
-<Seo {title} {metadescription} />
-
-<div class="flex flex-col items-center justify-center">
-  <div class="mb-4">
-    <div class="justify-self-center">
-      <Toggle
-        id="toggle"
-        bind:checked
-        labelLeft="EddieHub"
-        labelRight="GitHub"
-        on:change={onChangeHandler}
-      />
+<div class="mt-12 flex flex-col items-center px-4">
+  <div class="space-y-8">
+    <h1
+      class="mx-auto bg-gradient-to-r from-skin-primary via-[#ff00ff] to-cyan-400 bg-clip-text text-center text-3xl font-bold text-transparent dark:to-[#e3ff00] md:text-6xl"
+    >
+      Good-first-issue-finder
+    </h1>
+    <h2 class="max-w-2xl text-center md:text-xl">
+      Your entry point for finding good first issues in the EddieHub Organization and in the
+      community
+    </h2>
+    <div class="flex justify-center gap-8">
+      <Link variant="primary" href="/app">Get started</Link>
+      <Link variant="secondary" href="/docs">Read docs</Link>
     </div>
   </div>
-  <Search bind:searchTerm={searchString} on:keyup={() => performSearch()} />
 </div>
-{#if $navigating}
-  <div class="mt-8 flex items-center justify-center gap-4">
-    <Loader background="off-background" /> Loading...
-  </div>
-{:else if intersectedArray.length > 0}
-  <div class="mb-8 flex flex-col items-center">
-    <Filter tags={githubData.labels} />
-  </div>
-  <div class="mb-4 space-y-4">
-    {#each intersectedArray as node}
-      <IssueCard issue={node.node} />
-    {/each}
-    {#if githubData.pageInfo.hasNextPage}
-      <div class="flex items-center justify-center">
-        <LoadMore isDisabled={loadDisabled} on:load={fetchMore} />
-      </div>
-    {/if}
-  </div>
-{:else}
-  <div class="text-center">Unfortunately, there were no issues found.</div>
-{/if}
